src/repository/pyproject.toml
[tool.poetry]
name = "repository"
version = "0.1.0"
description = ""
authors = ["bjoluc <mail@bjoluc.de>"]

[tool.poetry.dependencies]
python = "^3.7"
eve = "^1.1.1"
config2 = "^0.3.2"
jsonschema = "^3.2.0"
gunicorn = "^20.0.4"
pyyaml = "^5.3.1"

[tool.poetry.scripts]
start = "repository.__main__:main"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.2"
black = "^19.10b0"
mongomock = "^3.19.0"
pytest-env = "^0.6.2"
pytest-watch = "^4.2.0"
pytest-mock = "^3.1.0"
pytest-voluptuous = "^1.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
